pyproject.toml
[tool.poetry]
name = "codtrackerbot"
version = "0.1.0"
description = ""
authors = ["Matthias Klettner <matthias.klettner@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pydantic = {extras = ["dotenv"], version = "^1.8.2"}
"discord.py" = "^1.7.3"
tabulate = "^0.8.9"
aiosqlite = "^0.17.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
